chore: sync API artifacts and maintainability gates
